travelhub-experiment-2/reservas/requirements.txt
Flask==3.1.3
pytz==2025.2
psycopg2-binary==2.9.10
requests==2.32.3
